imdb-sentiment-analysis/requirements.txt
pytorch
transformers
datasets
scikit-learn
pandas